# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vault API",
    "version": "1.0.0",
    "description": "Document storage, workspaces and revocable share links. JSON over HTTPS with a session cookie. Every error has the shape { error: { code, message } }. The same routes are served under /api/v1."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "fs_session",
        "description": "Session cookie set by /api/auth/login or /api/auth/register."
      }
    },
    "schemas": {
...
      "ShareSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdBy": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-17T10:15:00.000Z"
          },
          "expiresAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "example": "2026-09-17T10:15:00.000Z"
          },
          "hasPassword": {
            "type": "boolean"
          },
          "maxDownloads": {
            "type": [
              "integer",
              "null"
            ]
          },
          "downloadCount": {
            "type": "integer"
          },
          "allowDownload": {
            "type": "boolean"
          },
          "allowedEmails": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "activity": {
            "$ref": "#/components/schemas/ShareActivity"
          }
        },
        "required": [
          "id",
          "createdBy",
          "createdAt",
          "expiresAt",
          "hasPassword",
          "maxDownloads",
          "downloadCount",
          "allowDownload",
          "allowedEmails",
          "activity"
        ],
        "additionalProperties": false
      },
...
    },
...
  },
  "paths": {
...
    "/api/documents/{id}/shares": {
      "get": {
        "summary": "A document's live links and their activity",
        "tags": [
          "Sharing"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "shares": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ShareSummary"
                      }
                    }
                  },
                  "required": [
                    "shares"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Not signed in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not a member (deliberately identical).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/shares": {
      "post": {
        "summary": "Create a share link",
        "tags": [
          "Sharing"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShareRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "share": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "url": {
                          "type": "string",
                          "format": "uri",
                          "description": "Contains the token. Returned only here, once."
                        },
                        "expiresAt": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "format": "date-time",
                          "example": "2026-09-17T10:15:00.000Z"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2026-09-17T10:15:00.000Z"
                        },
                        "hasPassword": {
                          "type": "boolean"
                        },
                        "maxDownloads": {
                          "type": [
                            "integer",
                            "null"
                          ]
                        },
                        "allowDownload": {
                          "type": "boolean"
                        },
                        "allowedEmails": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      },
                      "required": [
                        "id",
                        "url",
                        "expiresAt",
                        "createdAt",
                        "hasPassword",
                        "maxDownloads",
                        "allowDownload",
                        "allowedEmails"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "share"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (`VALIDATION_FAILED` with details, or a specific code).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Signed in and a member, but the role does not allow this.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not a member (deliberately identical).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/shares/{id}": {
      "patch": {
        "summary": "Edit a link's expiry, password or download limit",
        "tags": [
          "Sharing"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateShareRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "share": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "expiresAt": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "format": "date-time",
                          "example": "2026-09-17T10:15:00.000Z"
                        },
                        "hasPassword": {
                          "type": "boolean"
                        },
                        "maxDownloads": {
                          "type": [
                            "integer",
                            "null"
                          ]
                        },
                        "downloadCount": {
                          "type": "integer"
                        },
                        "allowDownload": {
                          "type": "boolean"
                        },
                        "allowedEmails": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      },
                      "required": [
                        "id",
                        "expiresAt",
                        "hasPassword",
                        "maxDownloads",
                        "downloadCount",
                        "allowDownload",
                        "allowedEmails"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "share"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (`VALIDATION_FAILED` with details, or a specific code).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Signed in and a member, but the role does not allow this.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not a member (deliberately identical).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Revoke a link",
        "tags": [
          "Sharing"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Success, no content"
          },
          "401": {
            "description": "Not signed in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Signed in and a member, but the role does not allow this.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not a member (deliberately identical).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/shares/{id}/events": {
      "get": {
        "summary": "A link's access history",
        "tags": [
          "Sharing"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "events": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "accessedAt": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2026-09-17T10:15:00.000Z"
                          },
                          "outcome": {
                            "$ref": "#/components/schemas/ShareOutcome"
                          },
                          "userAgent": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "viewer": {
                            "type": "string",
                            "pattern": "^[0-9a-f]{8}$",
                            "description": "Opaque marker for \"the same viewer\". Never an address."
                          },
                          "email": {
                            "type": [
                              "string",
                              "null"
                            ],
                            "description": "The address the viewer proved, on a link restricted to named people."
                          }
                        },
                        "required": [
                          "accessedAt",
                          "outcome",
                          "userAgent",
                          "viewer",
                          "email"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "events"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Not signed in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not a member (deliberately identical).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/shares/{token}": {
      "get": {
        "summary": "Resolve a link",
        "description": "Records nothing. A locked link reveals only that a password is required.",
        "tags": [
          "Public sharing"
        ],
        "security": [],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 10,
              "maxLength": 200
            },
            "required": true,
            "name": "token",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicShare"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not a member (deliberately identical).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "Gone: revoked, expired, used up, or already used.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited or locked out. See Retry-After.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/shares/{token}/unlock": {
      "post": {
        "summary": "Enter a link password",
        "description": "Sets an HttpOnly cookie that unlocks this link for one hour.",
        "tags": [
          "Public sharing"
        ],
        "security": [],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 10,
              "maxLength": 200
            },
            "required": true,
            "name": "token",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnlockRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Success, no content"
          },
          "400": {
            "description": "Invalid request (`VALIDATION_FAILED` with details, or a specific code).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not a member (deliberately identical).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "Gone: revoked, expired, used up, or already used.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited or locked out. See Retry-After.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/shares/{token}/code": {
      "post": {
        "summary": "Email a one-time code to open a link restricted to named people",
        "description": "Always 202, whether or not the address is on the link. A code is sent only to addresses on the link, at most 3 per 15 minutes.",
        "tags": [
          "Public sharing"
        ],
        "security": [],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 10,
              "maxLength": 200
            },
            "required": true,
            "name": "token",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareCodeRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (`VALIDATION_FAILED` with details, or a specific code).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not a member (deliberately identical).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "Gone: revoked, expired, used up, or already used.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited or locked out. See Retry-After.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/shares/{token}/verify": {
      "post": {
        "summary": "Enter the one-time code",
        "description": "Sets an HttpOnly cookie naming the verified address, for one hour. Only the newest code works, for 10 minutes and 5 tries.",
        "tags": [
          "Public sharing"
        ],
        "security": [],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 10,
              "maxLength": 200
            },
            "required": true,
            "name": "token",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareCodeVerifyRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Success, no content"
          },
          "400": {
            "description": "Invalid request (`VALIDATION_FAILED` with details, or a specific code).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not a member (deliberately identical).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "Gone: revoked, expired, used up, or already used.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited or locked out. See Retry-After.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/shares/{token}/view": {
      "post": {
        "summary": "Record a page view",
        "tags": [
          "Public sharing"
        ],
        "security": [],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 10,
              "maxLength": 200
            },
            "required": true,
            "name": "token",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Success, no content"
          },
          "401": {
            "description": "Not signed in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not a member (deliberately identical).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "410": {
            "description": "Gone: revoked, expired, used up, or already used.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited or locked out. See Retry-After.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}